[v0.87][maintainability] Split learning_export.rs by responsibility
